+++ src/core/xml_parser.c
@@ -44,7 +44,7 @@ int xml_parser_compare_multiple(const char *string, const char *match)
     const char *next;
     do {
         next = strchr(string, '|');
-        size_t length = next ? (next - string) : strlen(string);
+        size_t length = next ? (size_t) (next - string) : strlen(string);
         if (strncmp(string, match, length) == 0) {
             return 1;
         }
