+++ src/window/editor/custom_messages.c
@@ -119,7 +119,7 @@ static void draw_foreground(void)
         y_offset += MESSAGES_ROW_HEIGHT;
     }
 
-    for (int i = 8; i < MAX_BUTTONS; i++) {
+    for (size_t i = 8; i < MAX_BUTTONS; i++) {
         large_label_draw(buttons[i].x, buttons[i].y, buttons[i].width / 16, data.focus_button_id == i + 1 ? 1 : 0);
     }
